pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pykworldsim"
version = "3.0.0"
description = "Production-grade ECS world simulation framework — people, cities, relationships, and more"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Kairav Dutta" }]
requires-python = ">=3.10"
keywords = ["simulation", "ecs", "entity-component-system", "world", "agent", "social-simulation"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.scripts]
pykworldsim = "pykworldsim.cli:app"

[project.entry-points."pykworldsim.systems"]
movement = "pykworldsim.core.systems.movement:MovementSystem"
physics  = "pykworldsim.core.systems.physics:PhysicsSystem"
social   = "pykworldsim.core.systems.social:SocialSystem"
event    = "pykworldsim.core.systems.event_system:EventSystem"

[project.urls]
Homepage   = "https://github.com/ka1rav6/pykworldsim"
Repository = "https://github.com/ka1rav6/pykworldsim"
Issues     = "https://github.com/ka1rav6/pykworldsim/issues"

[tool.setuptools.packages.find]
where   = ["."]
include = ["pykworldsim*"]

[tool.setuptools.package-data]
pykworldsim = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.mypy]
python_version         = "3.10"
strict                 = true
ignore_missing_imports = true

[tool.ruff]
line-length    = 88
target-version = "py310"
